+++ launcher-app/src/main/java/dk/aau/cs/giraf/launcher/settings/SettingsActivity.java
@@ -9,13 +9,15 @@ import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.provider.Settings;
-import android.support.v4.app.FragmentActivity;
 import android.view.View;
 import android.widget.AdapterView;
 
 import java.util.ArrayList;
 
 import dk.aau.cs.giraf.activity.GirafActivity;
+import dk.aau.cs.giraf.dblib.Helper;
+import dk.aau.cs.giraf.dblib.controllers.ProfileController;
+import dk.aau.cs.giraf.dblib.models.Profile;
 import dk.aau.cs.giraf.gui.GProfileSelector;
 import dk.aau.cs.giraf.gui.GirafButton;
 import dk.aau.cs.giraf.launcher.R;
@@ -23,14 +25,11 @@ import dk.aau.cs.giraf.launcher.helper.Constants;
 import dk.aau.cs.giraf.launcher.helper.LauncherUtility;
 import dk.aau.cs.giraf.launcher.settings.settingsappmanagement.AppManagementFragment;
 import dk.aau.cs.giraf.launcher.settings.settingsappmanagement.AppsFragmentInterface;
-import dk.aau.cs.giraf.dblib.Helper;
-import dk.aau.cs.giraf.dblib.controllers.ProfileController;
-import dk.aau.cs.giraf.dblib.models.Profile;
 
 /**
-* Activity responsible for handling Launcher settings_activity and starting
-* other setting-related activities.
-*/
+ * Activity responsible for handling Launcher settings_activity and starting
+ * other setting-related activities.
+ */
 public class SettingsActivity extends GirafActivity
         implements SettingsListFragment.SettingsListFragmentListener,
         AppsFragmentInterface {
@@ -83,28 +82,25 @@ public class SettingsActivity extends GirafActivity
             }
         });
 
-        addGirafButtonToActionBar(changeUserButton,LEFT);
+        addGirafButtonToActionBar(changeUserButton, LEFT);
 
         final Helper mHelper = LauncherUtility.getOasisHelper(this);
 
-        final int childId = getIntent().getExtras().getInt(Constants.CHILD_ID);
-        final int guardianId = getIntent().getExtras().getInt(Constants.GUARDIAN_ID);
+        final long childId = getIntent().getExtras().getLong(Constants.CHILD_ID);
+        final long guardianId = getIntent().getExtras().getLong(Constants.GUARDIAN_ID);
 
         mLoggedInGuardian = mHelper.profilesHelper.getProfileById(guardianId);
 
-        if(childId != Constants.NO_CHILD_SELECTED_ID)
-        {
+        if (childId != Constants.NO_CHILD_SELECTED_ID) {
             mCurrentUser = mHelper.profilesHelper.getProfileById(childId);
             mProfileSelector = new GProfileSelector(this, mLoggedInGuardian, null);
-        }
-        else
-        {
+        } else {
             mCurrentUser = mHelper.profilesHelper.getProfileById(guardianId);
             mProfileSelector = new GProfileSelector(this, mLoggedInGuardian, mCurrentUser);
         }
 
         // Change the title of the action bar to include the name of the current user
-        if(mCurrentUser != null) {
+        if (mCurrentUser != null) {
             this.setActionBarTitle(getString(R.string.settingsFor) + mCurrentUser.getName());
         }
 
@@ -138,15 +134,15 @@ public class SettingsActivity extends GirafActivity
     /**
      * This is used to set the onClickListener for a new ProfileSelector
      * It must be used everytime a new selector is set.
-     * */
-    private void setProfileSelectorClickListener(){
+     */
+    private void setProfileSelectorClickListener() {
         // Handles a selection returned from the profile select dialog
         mProfileSelector.setOnListItemClick(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
                 ProfileController pc = new ProfileController(SettingsActivity.this);
                 // Get a profile from id returned from dialog
-                mCurrentUser = pc.getProfileById((int) id);
+                mCurrentUser = pc.getProfileById(id);
                 // Close it when a selection has been made
                 mProfileSelector.dismiss();
 
@@ -386,8 +382,7 @@ public class SettingsActivity extends GirafActivity
         // Only add new transaction if the user clicked a non-active fragment
         if (mActiveFragment == null || !mActiveFragment.equals(fragment)) {
 
-            if(mActiveSupportFragment != null)
-            {
+            if (mActiveSupportFragment != null) {
                 android.support.v4.app.FragmentTransaction ft = mSupportFragManager.beginTransaction();
                 ft.remove(mActiveSupportFragment);
                 ft.commit();
@@ -413,10 +408,8 @@ public class SettingsActivity extends GirafActivity
     public void setActiveFragment(android.support.v4.app.Fragment fragment) {
 
         // Only add new transaction if the user clicked a non-active fragment
-        if (mActiveSupportFragment == null || !mActiveSupportFragment.equals(fragment))
-        {
-            if(mActiveFragment != null)
-            {
+        if (mActiveSupportFragment == null || !mActiveSupportFragment.equals(fragment)) {
+            if (mActiveFragment != null) {
                 FragmentTransaction ft = mFragManager.beginTransaction();
                 ft.remove(mActiveFragment);
                 ft.commit();
@@ -467,6 +460,7 @@ public class SettingsActivity extends GirafActivity
 
     /**
      * gets the current user
+     *
      * @return
      */
     public Profile getCurrentUser() {
@@ -475,6 +469,7 @@ public class SettingsActivity extends GirafActivity
 
     /**
      * get the current logged in guardian
+     *
      * @return
      */
     @Override
